refactor: unify corecrypto client public key impls in a common module
# Conflicts:
#	crypto-ffi/src/generic/mod.rs
#	crypto-ffi/src/wasm/mod.rs
