+++ crypto-ffi/src/core_crypto/mod.rs
@@ -1,3 +1,4 @@
+mod client;
 pub mod logger;
 pub mod mls_transport;
